changed user impact one-off job into continuous job
